differentiate between mode and routingMode
Not sure if this holds up, but it works so far for the bushfire evacuation tests.  The idea is that one can do

  tripRouter.addModule( routingMode, new NetworkRouter( mode ) ) ;

i.e. the mode through which it is called is different (= more specialized: carFreespeed, carCongested, ...)
from the mode that the leg in the end will have.

I can see that this is not terribly expressive.  But it works without changes on the outermost data model.