locationchoice work
I replaced many lines of code by more central constructs.  The results are NOT bit-compatible to what I had before.  I can't even say that the distribution looks the same.  It is, however, clearly declining with distance, so it is not random.  It is just not flat in logarithmic bins, so it is not exp(-dist) either.  It wasn't that before, either.  I can't say if it was "more" correct before; however, now it is based on more central construct re plans info, time info, and routing.  We might also need iterations to sort some of this out, which we don't have in the current testing setup.