remove lines of code that were already commented out
I was wrong in my assessment earlier.  exp(-x) should look straight when the y-axis would be logarithmic.  (Logarithmic bins on the x-axis tests for x^-1, which we don't have here.)

And, plotting distance distribution on a log-linear plot, we are getting a nice straight line.

So everything is working as it should.