Try to repair the ActivityHandler setup.
There is now one ActivityEngine, which is called last, a bit similar to the default TeleportationEngine, which is also the departure handler that is called last.

With the departure handler, this is not necessarily the best setup, since we had fall-throughs where we would rather have an error.  However, other than with the departure handlers, which depend on modes, for activities I don't even know what to base the case selection on (e.g. in DynActHandler, this is based on an agent instanceof expression).  So it has to be fall-through for the time being.