update BicycleTest reference files, change in MatsimRandom calls
prior to routingMode ModeStatsControlerListener called
tripRouterProvider.get() to obatin a MainModeIdentifier, this caused the
tripRouterProvider in the bicycle contrib to create a new
BicycleTravelDisutility and call MatsimRandom.getLocalInstance to do
that, which moved the random number generator forward. With routingMode
this no longer happens, because ModeStatsControlerListener does not call
tripRouterProvider.get() any longer. The new input files were obtained
by calling MatsimRandom.reset() on the old matsim master version, so the
random number generator did not move forward. That result passed the
test in routingMode (besides the attribute routingMode being added in
plans file). Thanks Kai N for finding this problem.