Make default params consistent between two different variants in PlanCalcScoreConfigGroup.  This also fixes the RunMultimodalExample problem.
There are two different ways to load params for the default subpop:
		<module name="planCalcScore">
		    <parameterset type="scoringParameters"> (*)
		        <parameterset type="modeParams">
and
		<module name="planCalcScore">
		    <parameterset type="modeParams">

In the existing implementation, the line (*) overwrite the default
params for the default subpopulation.  This failed RunExamples that
used to work (and which were not secured by a regression test).  I
have now added the same default params to all execution paths.  I know
that in general we discourage default settings, but I have found that
they lead to standardization between less experienced users, which is
a good thing.