introduce detailed emissions table consistency check; pull apart the overly long and intertwined TestWarmEmissionAnalysisModule into 6 different test classes
The issue that triggered this here was that it was not clear how to
handle cases where the freeflow lookup failed in the detailed table
but the stopgo lookup worked, or the other way around.

The old approach was that it fell back to average entries for both lookups.

With the present approach, that would have been difficult -- it would
have been necessary to first test both looksups, and if one of them
fails, then explicitly fall back onto the average lookup.

So the decision was to avoid such situations completely; all detailed
tables need to contain entries for all traffic situations.  There is
now a consistency check early in the program flow.

In consequence, the test cases testing the old behavior (2 and 3) now
just have "exception" as expected behavior.