fix a bug in QueueWithBuffer: ...
when a link has send some vehicles but not all its vehicles from the
buffer in the previous time step (because the downstream link got full),
it was not able to fill up the buffer in the old implementation. With
the new implementation/ bug fix the size of the vehicles in the buffer
is subtracted from the flow capacity accumulation value per time step
such that the remaining size is still able to enter the buffer, i.e.
fill it up.
The test NodeTransitionTest accounts for this issue and fails if
resulting throughput values of links are wrong.