Carsharing patch (#1399)
* fixing a bug

the one way vehicle mapping to station should only be removed after the
agent enters the vehicle.

This was causing problems for parking space availability

* fixing a bug

parking space was not updated previously for one way station
car-sharing. Now when the agent enters the vehicle at the station the
parking is made available for the next user.