taxi: pre-compute paths when occupied drives are unknown on request submission
In some configurations, the taxi optimiser may not know the destination until the passenger is picked up. But we can obtain this information already on request submission and use it to pre-compute the origin-destination path in the background.

If sufficiently many threads are used, all such paths are found before the pickup simulation ends. Consequently, QSim is not (directly) slowed down by computing these paths.

4-6 threads is usually enough. It's recommended to specify a higher number than that if possible (see the new param in the taxi config) - of course, threads will probably be not 100% busy. By default, it is set to the number of available cores.