Renaming to avoid some confusion (#1736)
The short trips are now renamed to normal trip (i.e. the trips within the city or region, which is the main type of the PT trips in the simulation).

The long trips are now renamed to long distance trip (i.e. intercity trips). Usually the ticket for those trips are different from the normal trips. Therefore a different fare parameter is used for those trips.