The motorized interaction was working in a way that it was generating MotorizedInteraction events for each LinkLeave event, which were then passed on to _every_ agent.  We know that that approach does not scale. With the new additionalScoreEvent scoring, it was rather straightforward to come up with a design that solves that issue further upstream in a much more lightweight way.  However, with that new approach I could not reconstruct the illustrative test that an artificial motorized interaction event was thrown regularly.  I therefore constructed a test that now has real cars on the same links, and bases interaction penalties on that.  Since the old test case did not rely on actual car interactions at all, I did not go for the exact same regression test before moving on.
I was then able to delete quite a lot of material since it is no longer needed.

This still needs to be cleaned up.  Which includes making the tests pass, i.e. copying their output files into the test input directory.makeBicycleLinkBasedScorePluggable

The car interaction is still somewhat primitive -- a vehicle leaving a link gets a penalty that is multiplied with the number of cars that are on the link at that moment.  Evidently, this could be improved, by counting the cars that actually overtake the bicycle.  Not very difficult ...