dvrp: fix non-deterministic PersonStuckEvent due to rejection
Due to race conditions, DefaultPassengerEngine may get notified about a rejection
before, during or after its doSimStep() finishes.
This means the rejection may be processed in the current or next time step.

To overcome this non-determinism, all rejections from the current time step are processed
in the next time step.