adapt tests to improved QuadTree
Reason the results are different: The new QuadTree is also more exact. When searching for the closest item in a QuadTree to a given coordinate, the old code calculated the euclidean distance between item and coordinate, while the new code calculates the square of the euclidean distance (essentially not performing the final Math.sqrt() in the Pythagorean formula). If two floating point values are nearly equal and only differ in the least significant digit, it could happen that the square-root of the values were actually identical due to the limit of significant digits. This resulted in the old QuadTree in the first item being found to be returned, while the new QuadTree might find the second item (which is just a very small amount closer to x/y) to be returned.

The aforementioned case was observed in the failing tests, e.g. where one item had a calculated distance of Math.sqrt(52643.632572290044), while the other had Math.sqrt(52643.632572290040). The old code returned the first one, while the new code returns the second one.